requirements.txt
plotly<4.14.3
streamlit<0.74.1
pandas<1.2.0
Pillow<8.1.0